+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased (on main branch only)
 
+- `list process` asynchronously
+
 ## 2024-05-06 v0.6.2
 
 - Make `task quick-add` `--content` flag optional (this feature was lost in the 0.6.0 API change)